requirements.txt
# Flask core
Flask==3.1.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==21.2.0
blinker==1.9.0

# Email
sendgrid==6.10.0

# HTTP + OAuth
requests==2.31.0
msal==1.26.0

# Azure
azure-functions==1.18.0
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0

sqlalchemy
psycopg2-binary  # if using PostgreSQL
# or
pyodbc           # if using SQL Server (only if needed)

